blottner_parsing no longer header only
